Added

+++ 10_Structs/LateralController/AdFunctions.c
@@ -80,9 +80,9 @@ void print_scene(const VehicleType *ego_vehicle, const NeighborVehiclesType *veh
     size_t center_idx = 0;
     size_t right_idx = 0;
 
-    const int32_t offset_m = 10;
+    const float offset_m = 10;
 
-    for (int32_t i = 100; i >= -100; i -= offset_m)
+    for (int32_t i = 100; i >= -100; i -= (int32_t)(offset_m))
     {
         const VehicleType *left_vehicle = &vehicles->vehicles_left_lane[left_idx];
         const VehicleType *center_vehicle = &vehicles->vehicles_center_lane[center_idx];
@@ -229,7 +229,7 @@ LaneAssociationType longitudinal_control(const NeighborVehiclesType *vehicles, V
     const float crash_distance = mps_to_kmh(ego_vehicle->speed_mps) / 4.0F;
 
     const float front_distance = front_vehicle->distance_m;
-    const float rear_distance = fabs(rear_vehicle->distance_m);
+    const float rear_distance = fabsf(rear_vehicle->distance_m);
 
     bool front_vehicle_too_close = false;
 
@@ -329,7 +329,7 @@ bool gap_is_valid(const VehicleType *front_vehicle, const VehicleType *rear_vehi
     const float minimal_distance = mps_to_kmh(ego_vehicle->speed_mps) / 5.0F;
 
     const float front_distance = front_vehicle->distance_m;
-    const float rear_distance = fabs(rear_vehicle->distance_m);
+    const float rear_distance = fabsf(rear_vehicle->distance_m);
 
     if ((front_distance > minimal_distance) && (rear_distance > minimal_distance))
     {
