+++ 8_PointerArrays/ArrayLib2/utils.c
@@ -91,7 +91,7 @@ void clear_console()
 void sleep_console(int ms)
 {
 #ifdef _WIN32
-    Sleep(ms);
+    Sleep((DWORD)(ms));
 #else
     usleep(ms * 1000);
 #endif
